refactor: centralize unicode and callback imports
