+++ core/security_validator.py
@@ -11,7 +11,7 @@ from typing import Dict, Any, List, Callable
 
 from config.constants import FileProcessingLimits
 
-from utils.unicode_utils import sanitize_unicode_input
+from core.unicode_processor import sanitize_unicode_input
 from dataclasses import dataclass
 from enum import Enum
 from .security_patterns import (
@@ -20,7 +20,7 @@ from .security_patterns import (
 )
 from security.sql_validator import SQLInjectionPrevention
 from security.validation_exceptions import ValidationError
-from services.data_processing.callback_controller import (
+from security_callback_controller import (
     emit_security_event,
     SecurityEvent,
 )
