+++ services/analytics_processing.py
@@ -8,7 +8,7 @@ from services import get_analytics_service
 from services.ai_suggestions import generate_column_suggestions
 from services.upload_data_service import get_uploaded_data
 from utils.preview_utils import serialize_dataframe_preview
-from utils.unicode_utils import sanitize_unicode_input, safe_format_number
+from core.unicode_processor import sanitize_unicode_input, safe_format_number
 
 logger = logging.getLogger(__name__)
