+++ tests/integration/test_unicode_handling_integration.py
@@ -1,6 +1,6 @@
 import pandas as pd
 from plugins.service_locator import PluginServiceLocator
-from utils.unicode_utils import UnicodeProcessor, sanitize_dataframe
+from core.unicode_processor import UnicodeProcessor, sanitize_dataframe
 
 handler = PluginServiceLocator.get_unicode_handler()
