+++ tests/test_security_comprehensive.py
@@ -20,7 +20,7 @@ class TestSecurityVulnerabilities:
 
     def test_unicode_surrogate_handling(self):
         """Test Unicode surrogate character handling."""
-        from utils.unicode_utils import sanitize_unicode_input
+        from core.unicode_processor import sanitize_unicode_input
 
         # Test lone surrogates
         lone_surrogate = "\ud800\ud801"  # Invalid surrogate pair
@@ -31,7 +31,7 @@ class TestSecurityVulnerabilities:
 
     def test_sanitize_unicode_input_ascii_fallback(self):
         """Ensure ASCII-safe text is returned on failure."""
-        from utils.unicode_utils import sanitize_unicode_input
+        from core.unicode_processor import sanitize_unicode_input
 
         class BadStr:
             def __str__(self) -> str:
